requirements.txt
flask
line-bot-sdk
gunicorn
google-generativeai
google-genai
pillow
requests
python-dotenv
